requirements.txt
python-dotenv
openai
flask
flask-smorest
flask-cors